Update from original

+++ src/AppBundle/Service/MenuService.php
@@ -2,6 +2,7 @@
 namespace AppBundle\Service;
 
 use Doctrine\ORM\EntityManager;
+use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Router;
 
 
@@ -9,11 +10,14 @@ class MenuService
 {
     private $em;
     private $router;
+    private $appService;
 
-    public function __construct(EntityManager $em,Router $router)
+    public function __construct(EntityManager $em,Router $router,AppService $appService)
     {
-        $this->em     = $em;
-        $this->router = $router;
+        $this->em           = $em;
+        $this->router       = $router;
+        $this->appService   = $appService;
+
     }
 
     public function cleaMenu($entity){
@@ -26,7 +30,8 @@ class MenuService
         foreach ($itemsMenu as $item)
             $ids[] = $item->getId();
 
-        $this->em->getConnection()->exec("DELETE FROM MenuItem WHERE id IN (".implode(',',$ids).") ");
+        if( count($ids) > 0 )
+            $this->em->getConnection()->exec("DELETE FROM MenuItem WHERE id IN (".implode(',',$ids).") ");
     }
 
     public function beforeUpdate($args){
@@ -58,6 +63,33 @@ class MenuService
         $this->em->flush();
     }
 
+    public function getMenuItems(){
+
+        $dql    = "SELECT m FROM AppBundle:MenuItem m WHERE m.parent IS NULL ORDER BY m.position ASC";
+        $query  = $this->em->createQuery($dql);
+        $site   = $this->appService->getSetting("site_nom");
+
+        if(APC_ENABLE)
+            $query->useResultCache(true,86400,$site."_menu");
+
+        return $query->getResult();
+    }
+
+    public function getMenuUrl(Request $request,$url,$env){
+
+        if(strpos($url,"#") !== false && $request->getPathInfo() != "/"){
+            $url = "/".$url;
+        }
+        if($env == "dev"){
+            $exploded = explode("/",$url);
+            if(count($exploded) > 1 ){
+                $exploded[0] = "/app_dev.php";
+                $url = implode("/",$exploded);
+            }
+        }
+
+        return $url;
+    }
 
 
 }
